v2 schema

+++ truncate.js
@@ -1,13 +1,19 @@
-const fs = require('fs');
 const dbFilePath = './db/db.sqlite';
+const {schema, samples} = require('./db/schema');
 
-if(fs.existsSync(dbFilePath)){
-  fs.unlinkSync(dbFilePath);
-}
+const fs = require('fs');
+
+//delete db
+if(fs.existsSync(dbFilePath))fs.unlinkSync(dbFilePath);
 fs.writeFileSync(dbFilePath, '', {encoding: 'utf8'});
 
-
 const Database  = require('better-sqlite3');
-const sql = fs.readFileSync('./db/cms.sql', 'utf8');
 const db = new Database(dbFilePath);
-db.exec(sql);
+const {parse, insert} = require('./parse-schema');
+db.exec(parse(schema));
+
+const inserts = insert(samples);
+db.exec(inserts);
+
+const codeGen = require('./parse-model');
+codeGen.exec(schema);
